fix 3 precision print for stats

+++ utils.cc
@@ -28,48 +28,48 @@ unsigned long tval_diff_to_usec(timeval tval_start, timeval tval_end) {
 }
 
 // util template functions
-float get_mean(unsigned long* data, int num_trial) {
+double get_mean(unsigned long* data, int num_trial) {
     // mean
-    float mean = 0.0;
+    double mean = 0.0;
     for (int i = 0; i < num_trial; ++i) {
-        mean += (float) data[i];
+        mean += (double) data[i];
     }
-    mean = mean / num_trial;
+    mean = mean / (double) num_trial;
     return mean;
 }
 
-float get_sd(unsigned long* data, int num_trial) {
+double get_sd(unsigned long* data, int num_trial) {
     // mean
-    float mean = get_mean(data, num_trial);
+    double mean = get_mean(data, num_trial);
     // std
-    float sum_deviation = 0.0;
+    double sum_deviation = 0.0;
     for (int i = 0; i < num_trial; ++i) {
-        sum_deviation += ((float) data[i] - mean) * ((float) data[i] - mean);
+        sum_deviation += ((double) data[i] - mean) * ((double) data[i] - mean);
     }
     return sqrt(sum_deviation / num_trial);
 }
 
-float get_min(unsigned long* data, int num_trial) {
-    float min_val = 0;
+double get_min(unsigned long* data, int num_trial) {
+    double min_val = 0;
     if (num_trial > 0) {
-        min_val = (float) data[0];
+        min_val = (double) data[0];
     }
     for (int i = 0; i < num_trial; ++i) {
-        if ((float) data[i] < min_val) {
-            min_val = data[i];
+        if ((double) data[i] < min_val) {
+            min_val = (double) data[i];
         }
     }
     return min_val;
 }
 
-float get_max(unsigned long* data, int num_trial) {
-    float max_val = 0;
+double get_max(unsigned long* data, int num_trial) {
+    double max_val = 0;
     if (max_val > 0) {
-        max_val = (float) data[0];
+        max_val = (double) data[0];
     }
     for (int i = 0; i < num_trial; ++i) {
-        if ((float) data[i] > max_val) {
-            max_val = data[i];
+        if ((double) data[i] > max_val) {
+            max_val = (double) data[i];
         }
     }
     return max_val;
@@ -77,15 +77,15 @@ float get_max(unsigned long* data, int num_trial) {
 
 // returns the trimmed num_trial
 int trim_outlier(unsigned long* data, int num_trial) {
-    float mean = get_mean(data, num_trial);
-    float sd = get_sd(data, num_trial);
-    float min_val = get_min(data, num_trial);
-    float max_bound = mean + 3 * sd;
+    double mean = get_mean(data, num_trial);
+    double sd = get_sd(data, num_trial);
+    double min_val = get_min(data, num_trial);
+    double max_bound = mean + 3 * sd;
     max_bound = std::min(max_bound, min_val * 2);
     max_bound = std::min(max_bound, mean * 2);
     int write_idx = 0;
     for (int i = 0; i < num_trial; i++) {
-        if ((float) data[i] <= max_bound && i > 5) {
+        if ((double) data[i] <= max_bound && i > 5) {
             data[write_idx] = data[i];
             write_idx++;
         }
@@ -95,11 +95,11 @@ int trim_outlier(unsigned long* data, int num_trial) {
 
 void print_stats(unsigned long* data, int num_trial, int num_iter,
                  int num_unroll) {
-    float mean = get_mean(data, num_trial);
-    float sd = get_sd(data, num_trial);
-    float min_val = get_min(data, num_trial);
-    float max_val = get_max(data, num_trial);
-    float mean_op = mean / (float) num_iter / (float) num_unroll;
+    double mean = get_mean(data, num_trial);
+    double sd = get_sd(data, num_trial);
+    double min_val = get_min(data, num_trial);
+    double max_val = get_max(data, num_trial);
+    double mean_op = mean / (double) num_iter / (double) num_unroll;
     std::cout.precision(3);
     std::cout << std::fixed;
     std::cout << "num_trial: " << num_trial << std::endl
@@ -116,10 +116,10 @@ void print_trimmed_mean_std(unsigned long* data, int num_trial, int num_iter,
                             int num_unroll) {
     // int trimmed_num_trial = trim_outlier(data, num_trial);
     trim_outlier(data, num_trial);
-    float mean = get_mean(data, num_trial);
-    float sd = get_sd(data, num_trial);
-    float mean_op = mean / (float) num_iter / (float) num_unroll;
-    float sd_op = sd / (float) num_iter / (float) num_unroll;
+    double mean = get_mean(data, num_trial);
+    double sd = get_sd(data, num_trial);
+    double mean_op = mean / (double) num_iter / (double) num_unroll;
+    double sd_op = sd / (double) num_iter / (double) num_unroll;
     std::cout << "mean_trial: " << mean << std::endl
               << "std_trial: " << sd << std::endl
               << "mean_op: " << mean_op << std::endl
